pre cleanup

+++ src/backend/imaging/upscale/detail/shiftlut_model_format.h
@@ -43,11 +43,16 @@ struct Dimensions {
 
 MMLTK_SHIFTLUT_FORMAT_HD constexpr Dimensions dimensions(TableFamily family) noexcept {
     switch (family) {
-        case TableFamily::Low: return {kChannels, kSpatialTaps, kLowDomain};
-        case TableFamily::Depthwise: return {kChannels, kSpatialTaps, kHighDomain};
-        case TableFamily::Pointwise: return {kChannels, kChannels, kHighDomain};
-        case TableFamily::Up: return {kOutputPhases, kChannels, kHighDomain};
-        case TableFamily::Shifts: return {kStages, kShiftAxes, kChannels};
+        case TableFamily::Low:
+            return {kChannels, kSpatialTaps, kLowDomain};
+        case TableFamily::Depthwise:
+            return {kChannels, kSpatialTaps, kHighDomain};
+        case TableFamily::Pointwise:
+            return {kChannels, kChannels, kHighDomain};
+        case TableFamily::Up:
+            return {kOutputPhases, kChannels, kHighDomain};
+        case TableFamily::Shifts:
+            return {kStages, kShiftAxes, kChannels};
     }
     return {};
 }
@@ -57,12 +62,16 @@ MMLTK_SHIFTLUT_FORMAT_HD constexpr std::size_t table_offset(TableFamily family, 
     const auto depthwise = dimensions(TableFamily::Depthwise).elements();
     const auto pair = depthwise + dimensions(TableFamily::Pointwise).elements();
     switch (family) {
-        case TableFamily::Low: return 0;
-        case TableFamily::Depthwise: return low + static_cast<std::size_t>(stage) * pair;
-        case TableFamily::Pointwise: return low + static_cast<std::size_t>(stage) * pair + depthwise;
-        case TableFamily::Up: return low + kStages * pair;
-        case TableFamily::Shifts: return low + kStages * pair + dimensions(TableFamily::Up).elements() +
-            static_cast<std::size_t>(stage) * kShiftAxes * kChannels;
+        case TableFamily::Low:
+            return 0;
+        case TableFamily::Depthwise:
+            return low + static_cast<std::size_t>(stage) * pair;
+        case TableFamily::Pointwise:
+            return low + static_cast<std::size_t>(stage) * pair + depthwise;
+        case TableFamily::Up:
+            return low + kStages * pair;
+        case TableFamily::Shifts:
+            return low + kStages * pair + dimensions(TableFamily::Up).elements() + static_cast<std::size_t>(stage) * kShiftAxes * kChannels;
     }
     return 0;
 }
@@ -79,8 +88,10 @@ inline constexpr std::size_t kTableBlocks = 1 + kStages * 2 + 2;
 constexpr TableBlock table_block(std::size_t index) noexcept {
     TableFamily family = TableFamily::Low;
     Stage stage = Stage::First;
-    if (index == kTableBlocks - 1) family = TableFamily::Shifts;
-    else if (index == kTableBlocks - 2) family = TableFamily::Up;
+    if (index == kTableBlocks - 1)
+        family = TableFamily::Shifts;
+    else if (index == kTableBlocks - 2)
+        family = TableFamily::Up;
     else if (index != 0) {
         family = (index - 1) % 2 == 0 ? TableFamily::Depthwise : TableFamily::Pointwise;
         stage = static_cast<Stage>((index - 1) / 2);
@@ -91,18 +102,24 @@ constexpr TableBlock table_block(std::size_t index) noexcept {
     } else if (family == TableFamily::Up) {
         result.name = {'U', 'P', '_', 'M', 'S', 'B'};
     } else {
-        result.name = {family == TableFamily::Pointwise ? 'P' : 'D', 'W',
-                       static_cast<char>('0' + static_cast<char>(stage)), '_',
-                       family == TableFamily::Low ? 'L' : 'M', 'S', 'B'};
+        result.name = {family == TableFamily::Pointwise ? 'P' : 'D',
+                       'W',
+                       static_cast<char>('0' + static_cast<char>(stage)),
+                       '_',
+                       family == TableFamily::Low ? 'L' : 'M',
+                       'S',
+                       'B'};
     }
     return result;
 }
 
-inline constexpr std::size_t kTableElements =
-    table_offset(TableFamily::Shifts) + dimensions(TableFamily::Shifts).elements();
+inline constexpr std::size_t kTableElements = table_offset(TableFamily::Shifts) + dimensions(TableFamily::Shifts).elements();
 inline constexpr std::size_t kTableBytes = kTableElements * sizeof(float);
-inline constexpr std::size_t kMaximumTilePixels = kMaximumTileExtent * kMaximumTileExtent;
-constexpr std::size_t scratch_elements(std::size_t pixels) noexcept { return kRotatedBatch * kChannels * pixels; }
+inline constexpr std::size_t kMaximumTilePixels =
+    static_cast<std::size_t>(kMaximumTileExtent) * static_cast<std::size_t>(kMaximumTileExtent);
+constexpr std::size_t scratch_elements(std::size_t pixels) noexcept {
+    return static_cast<std::size_t>(kRotatedBatch) * static_cast<std::size_t>(kChannels) * pixels;
+}
 inline constexpr std::size_t kScratchElements = scratch_elements(kMaximumTilePixels);
 constexpr std::size_t decision_elements(std::size_t pixels) noexcept { return kDecisionCheckpoints * scratch_elements(pixels); }
 constexpr std::size_t decision_offset(Stage stage, Decision decision, std::size_t pixels) noexcept {
